pyproject.toml
[tool.poetry]
name = "djangoapp"
version = "0.1.0"
description = ""
authors = ["Evgeniy"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.9"
psycopg2-binary = "^2.9.9"
pillow = "^10.2.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
pytest-django = "^4.8.0"
pytest-factoryboy = "^2.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
